Moving AvatarLabel
