+++ amundsen_application/static/js/components/AvatarLabel/index.spec.tsx
