+++ amundsen_application/static/js/components/AvatarLabel/index.tsx
