+++ amundsen_application/static/js/components/AvatarLabel/styles.scss
