+++ amundsen_application/static/js/pages/TableDetailPage/LineageLink/index.tsx
@@ -3,7 +3,7 @@
 
 import * as React from 'react';
 
-import AvatarLabel from 'components/common/AvatarLabel';
+import AvatarLabel from 'components/AvatarLabel';
 import AppConfig from 'config/config';
 import { logClick } from 'ducks/utilMethods';
 import { TableMetadata } from 'interfaces/TableMetadata';
