+++ amundsen_application/static/js/pages/TableDetailPage/SourceLink/index.spec.tsx
@@ -4,7 +4,7 @@
 import * as React from 'react';
 import { mount } from 'enzyme';
 
-import AvatarLabel from 'components/common/AvatarLabel';
+import AvatarLabel from 'components/AvatarLabel';
 import AppConfig from 'config/config';
 import { ResourceType } from 'interfaces/Resources';
 import SourceLink, { SourceLinkProps } from '.';
