+++ amundsen_application/static/js/pages/TableDetailPage/WriterLink/index.tsx
@@ -4,7 +4,7 @@
 import * as React from 'react';
 import { OverlayTrigger, Popover } from 'react-bootstrap';
 
-import AvatarLabel from 'components/common/AvatarLabel';
+import AvatarLabel from 'components/AvatarLabel';
 import { logClick } from 'ducks/utilMethods';
 import { TableWriter } from 'interfaces';
